+++ requests_magic/__init__.py
@@ -6,4 +6,6 @@ from .item import Item
 from .request import Request
 from .pipeline import Pipeline, SimpleFilePipeline, SimpleConsolePipeline
 from .spider import Spider
-from .logger import Logger, set_logger_level
+from .mmlog import logger, console_handler
+
+