+++ requests_magic/exception.py
@@ -1,0 +1,2 @@
+class ExistingIdentityError(Exception):
+    pass